+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [1.144.0]
+
+### Added
+- Button: Added event "click(visualisation-id, download(sid))"
+
+### Fixed
+- Waterfall: Fixed resize after change tab
+
 ## [1.143.0]
 
 ### Added